Improvements to the MakeInFile script:
verify that the file identified in the data catalog truly does exist.  In the case of data from the processed folder, verify that the corresponding masked data truly does exist.

Make the script record the output path at entry creation; this allows us to mix output paths for different sets of runs, without interfering with MakeOutDirs.sh.

Write the length (number of entries) of the longest file in a job to LengthOfJob.txt, so that the duration of a job can be adjusted where appropriate.